КАО#R2: agent-team audit round 2 — fix 5 serious + 5 minor + 1 suggestion
Full-app КАО cycle (8 auditors → adversarial verify → fix → independent
re-verify): 11 confirmed findings, all fixed, 0 regressions, round closed
at 0/0/0/0. Live-browser demo checks green.

Serious:
- КАО#R2-01 StyledToast: toasts invisible to screen readers (toast.custom
  bypasses ariaProps; WCAG 4.1.3) → role/aria-live on the root div.
- КАО#R2-02 visual_review route: GET 500 on half-step score
  (Decimal('7.5') vs _ScoreResponse.score: int) → float.
- КАО#R2-03 visual_review service: int(row.score) truncated half-steps in
  winner selection (7.5 vs 7.0 → spurious tie) → float.
- КАО#R2-04 orchestrator: enter_awaiting_visual_review lacked on_resume —
  24h auto-finalize + 1h vision auto-resume never advanced sessions.
  Wired _vr_on_resume (fresh AsyncSessionLocal, exception-guarded);
  main.py stale comment corrected (cross-restart re-arm remains VR-21).
- КАО#R2-05 useTimelinePlayer: Restart while playing soft-locked the
  player (stale currentChapterIdx in rAF closure) → chapter index
  mirrored into a ref (setChapterIdx) read by the tick.

Minor:
- КАО#R2-06 WSStatusPill overlapped MetricsPanel → panel nudges down.
- КАО#R2-07(+b) single-panel invariant: View Result / Intervene buttons,
  'c'/'i' shortcuts and CompletionBanner.onViewResult (found by the
  re-verify round) now route through switchToPanel().
- КАО#R2-08 sandbox-tab/frame: preview iframes get title= (WCAG 4.1.2).
- КАО#R2-09 react-router-dom ^6.30.4 (GHSA-2j2x-hqr9-3h42 open redirect);
  npm audit back to 0.
- КАО#R2-10 pyproject.toml: google-generativeai → google-genai (matches
  actual imports), python-jose >=3.4.0,<4 (CVE-2024-33664).

Suggestion:
- КАО#R2-11 WhatNextCta: passive overlay mis-declared role="dialog" →
  role="region" + aria-live.

Tests: new backend/tests/test_kao_r2_fixes.py; integrity sweep green
(tsc, npm audit 0, py_compile, in-process suites).

+++ backend/app/api/routes/visual_review.py
@@ -92,7 +92,10 @@ class _ScreenshotResponse(BaseModel):
 
 
 class _ScoreResponse(BaseModel):
-    score: int
+    # КАО#R2-02 — column is Numeric(3,1) (VR-27 half-steps) → SQLAlchemy yields
+    # Decimal('7.5'); an int field rejects fractional values and FastAPI turns the
+    # response_model ValidationError into a 500. Must be float to match _ScoreItem.
+    score: float
     source: str
     submitted_by: str | None = None
